refactor♻️: Refatorado o controllers de Authentication + Interface + Tests
